Python group
So we showed up (about five of us) to this group and found out no one else showed up, and instead decided to do our own work. It was for a Python group. So I ended up pairing up with one guy and making a game (well, starting to) that weighs items in kgs or lbs and asks for inputs, gives outputs, etc. This is a WIP but I should be able to work on it in the next few days. I was using a REPL browser ide-thing with Python 2.7 on it